+++ src/pages/RecruitmentPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import UserRecruitment from '@components/recruitment/user-recruitment/UserRecruitment'
-import RecruitmentList from '@components/recruitment/RecruitmentList'
+import RecruitmentList from '@src/components/recruitment/recruitment-list/RecruitmentList'
 import RecruitmentHeader from '@src/components/recruitment/recruitment-header/RecruitmentHeader'
 import SearchFilterBar from '@src/components/recruitment/srearch-filter-bar/SearchFilterBar'
